+++ axum-live-view/Cargo.toml
@@ -28,7 +28,6 @@ tower = { version = "0.4", default-features = false, features = ["util"] }
 tower-layer = "0.3"
 tower-service = "0.3"
 tracing = "0.1"
-uuid = { version = "0.8", features = ["v4", "serde"] }
 
 [dev-dependencies]
 tokio = { version = "1.14", features = ["full"] }
